+++ huddle_cluster_pkg/__init__.py
@@ -54,11 +54,13 @@ from huddle_cluster_pkg.cluster_service_discovery import ServiceDiscovery
 from huddle_cluster_pkg.cluster_ha                import ClusterHA
 from huddle_cluster_pkg.cluster_multi_region      import MultiRegionManager
 from huddle_cluster_pkg.cluster_circuit_breaker   import ClusterCircuitBreaker
-from huddle_cluster_pkg.cluster_rate_limiter      import ClusterRateLimiter
+from huddle_cluster_pkg.cluster_rate_limiter        import ClusterRateLimiter
+from huddle_cluster_pkg.cluster_canary_deployment   import ClusterCanaryDeployment
 
 __all__ = [
     "MasterNode", "NodeRecord", "AgentNode",
     "ClusterScheduler", "ClusterAutoScaler",
     "ClusterRollingUpdater", "ServiceDiscovery", "ClusterHA",
-    "MultiRegionManager", "ClusterCircuitBreaker", "ClusterRateLimiter",
+    "MultiRegionManager", "ClusterCircuitBreaker",
+    "ClusterRateLimiter", "ClusterCanaryDeployment",
 ]